\name{behaviour}
\alias{behaviour}
\docType{data}
\title{Binary three way data array}
\description{The data contains the reactions of a group of people to certain situations}
\usage{data(behaviour)}
\format{
A data array with a hipotetical 128 persons x 14 emotions x 11 situations:
}
\examples{
data(behaviour)
str(behaviour)
}
\keyword{datasets}
