% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryPARAFAC.R
\name{BinaryPARAFAC}
\alias{BinaryPARAFAC}
\title{Perform binary PARAFAC analysis on a three-dimensional dataset.}
\usage{
BinaryPARAFAC(
  X,
  s = 2,
  I,
  J,
  K,
  tolerance = 1e-04,
  penalization = 0.2,
  num_max_iters = 100,
  OptimMethod = "CG",
  Nombres = NULL
)
}
\arguments{
\item{X}{The three-dimensional data matrix.}

\item{s}{The number of components to extract (default: 2).}

\item{I}{The number of levels in the first dimension of the data.}

\item{J}{The number of levels in the second dimension of the data.}

\item{K}{The number of levels in the third dimension of the data.}

\item{tolerance}{The tolerance for algorithm convergence (default: 1e-04).}

\item{penalization}{The value of the penalty to apply (default: 0.2).}

\item{num_max_iters}{The maximum number of allowed iterations (default: 100).}

\item{OptimMethod}{The optimization method to use (default: "CG").}

\item{Nombres}{The names of variables in the data matrix (optional).}
}
\value{
A list containing the results of binary PARAFAC analysis.
}
\description{
This function implements binary PARAFAC analysis on a three-dimensional dataset,
using the user-specified algorithm.
}
\examples{
BinaryPARAFAC(X = behaviour, s = 2, I = 128, J = 14, K = 11)

}
